requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Streamlit and Web Interface
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# LangChain and LLM Integration
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-community>=0.0.10
python-dotenv>=1.0.0

# MLOps and Experiment Tracking
mlflow>=2.8.0
wandb>=0.15.0
tensorboard>=2.13.0

# Image Processing and Visualization
openslide-python>=1.3.0
tifffile>=2023.7.0
imageio>=2.31.0
albumentations>=1.3.0

# Data Handling
pandas>=2.0.0
h5py>=3.9.0
pyyaml>=6.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0
colorama>=0.4.6

# Development and Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Additional Dependencies
requests>=2.31.0
aiohttp>=3.8.0
asyncio>=3.4.3 